Backtrack for small holes in grid.
When passing a point where the current piece cannot be placed, bail
entirely if the area reachable from that point is smaller than _any_
of the possible pieces to avoid wasting time on unsolveable states.

In the underconstrained case (less than 3 pre-filled slots) this may
cause backtracking that skips solutions, but it is worth it to get
much more speed for the fully constrained case.